+++ src/compas/datastructures/mesh/mesh.py
@@ -639,14 +639,14 @@ class Mesh(HalfEdge):
         dy = dy or dx
         ny = ny or nx
 
-        V, U = meshgrid(linspace(0, dx, nx + 1), linspace(0, dy, ny + 1), indexing='ij')
+        U, V = meshgrid(linspace(0, dx, nx + 1), linspace(0, dy, ny + 1), indexing='ij')
 
         quads = [[
-            [U[i + 0][j + 0], V[i + 0][j + 0], 0],
-            [U[i + 0][j + 1], V[i + 0][j + 1], 0],
-            [U[i + 1][j + 1], V[i + 1][j + 1], 0],
-            [U[i + 1][j + 0], V[i + 1][j + 0], 0]
-        ] for i, j in product(range(ny), range(nx))]
+            [U[i + 0][j + 0], V[i + 0][j + 0], 0.0],
+            [U[i + 0][j + 1], V[i + 0][j + 1], 0.0],
+            [U[i + 1][j + 1], V[i + 1][j + 1], 0.0],
+            [U[i + 1][j + 0], V[i + 1][j + 0], 0.0]
+        ] for i, j in product(range(nx), range(ny))]
 
         return cls.from_polygons(quads)