+++ src/compas_ghpython/__init__.py
@@ -22,7 +22,7 @@ if compas.RHINO:
     from .utilities import *  # noqa: F401 F403
 
 
-__version__ = '1.9.1'
+__version__ = '1.10.0'
 
 
 def get_grasshopper_plugin_path(version):