+++ src/compas_rhino/geometry/__init__.py
@@ -9,104 +9,65 @@ geometry
 
 Wrappers for Rhino objects that can be used to convert Rhino geometry and data to COMPAS objects.
 
-.. code-block:: python
+Classes
+=======
 
-    import compas_rhino
-    from compas_rhino.geometry import RhinoMesh
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
 
-    guid = compas_rhino.select_mesh()
-    mesh = RhinoMesh.from_guid(guid).to_compas()
-
-
-BaseRhinoGeometry
-=================
-
-.. autoclass:: BaseRhinoGeometry
-    :members: from_geometry, from_selection, to_compas, from_guid, from_object, transform
-
-----
-
-RhinoPoint
-==========
-
-.. autoclass:: RhinoPoint
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoVector
-===========
-
-.. autoclass:: RhinoVector
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoLine
-=========
-
-.. autoclass:: RhinoLine
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoPlane
-==========
-
-.. autoclass:: RhinoPlane
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoMesh
-=========
-
-.. autoclass:: RhinoMesh
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoCurve
-==========
-
-.. autoclass:: RhinoCurve
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
-
-----
-
-RhinoSurface
-============
-
-.. autoclass:: RhinoSurface
-    :members: from_geometry, from_selection, to_compas
-    :no-show-inheritance:
+    RhinoGeometry
+    RhinoBox
+    RhinoCircle
+    RhinoCone
+    RhinoCurve
+    RhinoCylinder
+    RhinoEllipse
+    RhinoLine
+    RhinoMesh
+    RhinoPlane
+    RhinoPoint
+    RhinoPolyline
+    RhinoSphere
+    RhinoSurface
+    RhinoVector
 
 """
 from __future__ import absolute_import
 
-from ._geometry import BaseRhinoGeometry
+from ._geometry import RhinoGeometry
 
+from .box import RhinoBox
+from .circle import RhinoCircle
+from .cone import RhinoCone
 from .curve import RhinoCurve
+from .cylinder import RhinoCylinder
+from .ellipse import RhinoEllipse
 from .line import RhinoLine
 from .mesh import RhinoMesh
 from .plane import RhinoPlane
 from .point import RhinoPoint
+from .polyline import RhinoPolyline
+from .sphere import RhinoSphere
 from .surface import RhinoSurface
 from .vector import RhinoVector
 
+BaseRhinoGeometry = RhinoGeometry
+
 __all__ = [
-    'BaseRhinoGeometry',
+    'RhinoGeometry',
+    'RhinoBox',
+    'RhinoCircle',
+    'RhinoCone',
     'RhinoCurve',
+    'RhinoCylinder',
+    'RhinoEllipse',
     'RhinoLine',
     'RhinoMesh',
     'RhinoPlane',
     'RhinoPoint',
+    'RhinoPolyline',
+    'RhinoSphere',
     'RhinoSurface',
-    'RhinoVector'
+    'RhinoVector',
 ]